+++ router/types/get_quote_request_test.go
@@ -1,9 +1,7 @@
 package types_test
 
 import (
-	"net/http"
 	"net/http/httptest"
-	"strings"
 	"testing"
 
 	"github.com/labstack/echo/v4"
@@ -21,9 +19,7 @@ func TestGetQuoteRequestUnmarshal(t *testing.T) {
 		name           string
 		queryParams    map[string]string
 		expectedResult *types.GetQuoteRequest
-		expectedError  error
-		expectedStatus int
-		expectedBody   string
+		expectedError  bool
 	}{
 		{
 			name: "valid request with tokenIn and tokenOut",
@@ -43,9 +39,6 @@ func TestGetQuoteRequestUnmarshal(t *testing.T) {
 				SingleRoute:    true,
 				ApplyExponents: true,
 			},
-			expectedError:  nil,
-			expectedStatus: http.StatusOK,
-			expectedBody:   "",
 		},
 		{
 			name: "invalid singleRoute param",
@@ -56,9 +49,7 @@ func TestGetQuoteRequestUnmarshal(t *testing.T) {
 				"applyExponents": "true",
 			},
 			expectedResult: nil,
-			expectedError:  nil,
-			expectedStatus: http.StatusInternalServerError,
-			expectedBody:   `{"message":"strconv.ParseBool: parsing \"invalid\": invalid syntax"}`,
+			expectedError:  true,
 		},
 		{
 			name: "invalid applyExponents param",
@@ -69,9 +60,7 @@ func TestGetQuoteRequestUnmarshal(t *testing.T) {
 				"applyExponents": "invalid",
 			},
 			expectedResult: nil,
-			expectedError:  nil,
-			expectedStatus: http.StatusInternalServerError,
-			expectedBody:   `{"message":"strconv.ParseBool: parsing \"invalid\": invalid syntax"}`,
+			expectedError:  true,
 		},
 		{
 			name: "invalid tokenIn param",
@@ -82,7 +71,7 @@ func TestGetQuoteRequestUnmarshal(t *testing.T) {
 				"applyExponents": "true",
 			},
 			expectedResult: nil,
-			expectedError:  types.ErrTokenInNotValid,
+			expectedError:  true,
 		},
 		{
 			name: "invalid tokenOut param",
@@ -93,7 +82,7 @@ func TestGetQuoteRequestUnmarshal(t *testing.T) {
 				"applyExponents": "true",
 			},
 			expectedResult: nil,
-			expectedError:  types.ErrTokenOutNotValid,
+			expectedError:  true,
 		},
 	}
 
@@ -112,20 +101,15 @@ func TestGetQuoteRequestUnmarshal(t *testing.T) {
 			var result types.GetQuoteRequest
 			err := (&result).UnmarshalHTTPRequest(c)
 
-			if tc.expectedError != nil {
+			if tc.expectedError {
 				assert.Error(t, err)
-				assert.Equal(t, tc.expectedError, err)
 				return
 			}
 
 			assert.NoError(t, err)
-			assert.Equal(t, tc.expectedStatus, rec.Code)
-			assert.Equal(t, tc.expectedBody, strings.TrimSpace(rec.Body.String())) // JSONEq fails
 
 			// GetQuoteRequest must contain the expected result if the status is OK
-			if tc.expectedStatus == http.StatusOK {
-				assert.Equal(t, tc.expectedResult, &result)
-			}
+			assert.Equal(t, tc.expectedResult, &result)
 		})
 	}
 }