+++ CHANGELOG.md
@@ -35,7 +35,16 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
-## Unreleased
+## v25.12.0
+
+- Add router and candidate route search options to disable route cache
+
+## v25.11.0
+
+- Fix concurrency bug in data fetcher prefetching.
+- Fix error response status codes in /quote and /custom-direct-quote
+
+## v25.10.0
 
 - Orderbook plugin
 - Change tick iteration exit condition in CL in orderbook from zero to smallest Dec (18 decimals).